requirements.txt
fastapi
uvicorn
manim
requests
aiofiles
python-dotenv
